Keep context lifecycle test access private to unit test
